move more stuff from cmd to package
