+++ format_test.go
@@ -1,4 +1,4 @@
-package main
+package grpcurl
 
 import (
 	"bytes"
@@ -12,12 +12,10 @@ import (
 	"github.com/golang/protobuf/ptypes/struct"
 	"github.com/jhump/protoreflect/desc"
 	"google.golang.org/grpc/metadata"
-
-	"github.com/fullstorydev/grpcurl"
 )
 
 func TestRequestFactory(t *testing.T) {
-	source, err := grpcurl.DescriptorSourceFromProtoSets("../../testing/example.protoset")
+	source, err := DescriptorSourceFromProtoSets("testing/example.protoset")
 	if err != nil {
 		t.Fatalf("failed to create descriptor source: %v", err)
 	}
@@ -28,42 +26,42 @@ func TestRequestFactory(t *testing.T) {
 	}
 
 	testCases := []struct {
-		format         string
+		format         Format
 		input          string
 		expectedOutput []proto.Message
 	}{
 		{
-			format: "json",
+			format: FormatJSON,
 			input:  "",
 		},
 		{
-			format:         "json",
+			format:         FormatJSON,
 			input:          messageAsJSON,
 			expectedOutput: []proto.Message{msg},
 		},
 		{
-			format:         "json",
+			format:         FormatJSON,
 			input:          messageAsJSON + messageAsJSON + messageAsJSON,
 			expectedOutput: []proto.Message{msg, msg, msg},
 		},
 		{
 			// unlike JSON, empty input yields one empty message (vs. zero messages)
-			format:         "text",
+			format:         FormatText,
 			input:          "",
 			expectedOutput: []proto.Message{&structpb.Value{}},
 		},
 		{
-			format:         "text",
+			format:         FormatText,
 			input:          messageAsText,
 			expectedOutput: []proto.Message{msg},
 		},
 		{
-			format:         "text",
+			format:         FormatText,
 			input:          messageAsText + string(textSeparatorChar),
 			expectedOutput: []proto.Message{msg, &structpb.Value{}},
 		},
 		{
-			format:         "text",
+			format:         FormatText,
 			input:          messageAsText + string(textSeparatorChar) + messageAsText + string(textSeparatorChar) + messageAsText,
 			expectedOutput: []proto.Message{msg, msg, msg},
 		},
@@ -71,11 +69,15 @@ func TestRequestFactory(t *testing.T) {
 
 	for i, tc := range testCases {
 		name := fmt.Sprintf("#%d, %s, %d message(s)", i+1, tc.format, len(tc.expectedOutput))
-		rf, _ := formatDetails(tc.format, source, false, strings.NewReader(tc.input))
+		rf, _, err := RequestParserAndFormatterFor(tc.format, source, false, false, strings.NewReader(tc.input))
+		if err != nil {
+			t.Errorf("Failed to create parser and formatter: %v", err)
+			continue
+		}
 		numReqs := 0
 		for {
 			var req structpb.Value
-			err := rf.next(&req)
+			err := rf.Next(&req)
 			if err == io.EOF {
 				break
 			} else if err != nil {
@@ -86,8 +88,8 @@ func TestRequestFactory(t *testing.T) {
 			}
 			numReqs++
 		}
-		if rf.numRequests() != numReqs {
-			t.Errorf("%s: factory reported wrong number of requests: expecting %d, got %d", name, numReqs, rf.numRequests())
+		if rf.NumRequests() != numReqs {
+			t.Errorf("%s: factory reported wrong number of requests: expecting %d, got %d", name, numReqs, rf.NumRequests())
 		}
 	}
 }
@@ -95,7 +97,7 @@ func TestRequestFactory(t *testing.T) {
 // Handler prints response data (and headers/trailers in verbose mode).
 // This verifies that we get the right output in both JSON and proto text modes.
 func TestHandler(t *testing.T) {
-	source, err := grpcurl.DescriptorSourceFromProtoSets("../../testing/example.protoset")
+	source, err := DescriptorSourceFromProtoSets("testing/example.protoset")
 	if err != nil {
 		t.Fatalf("failed to create descriptor source: %v", err)
 	}
@@ -116,7 +118,7 @@ func TestHandler(t *testing.T) {
 		t.Fatalf("failed to create response message: %v", err)
 	}
 
-	for _, format := range []string{"json", "text"} {
+	for _, format := range []Format{FormatJSON, FormatText} {
 		for _, numMessages := range []int{1, 3} {
 			for _, verbose := range []bool{true, false} {
 				name := fmt.Sprintf("%s, %d message(s)", format, numMessages)
@@ -124,15 +126,14 @@ func TestHandler(t *testing.T) {
 					name += ", verbose"
 				}
 
-				_, formatter := formatDetails(format, source, verbose, nil)
+				_, formatter, err := RequestParserAndFormatterFor(format, source, false, !verbose, nil)
+				if err != nil {
+					t.Errorf("Failed to create parser and formatter: %v", err)
+					continue
+				}
 
 				var buf bytes.Buffer
-				h := handler{
-					out:        &buf,
-					descSource: source,
-					verbose:    verbose,
-					formatter:  formatter,
-				}
+				h := NewDefaultEventHandler(&buf, source, formatter, verbose)
 
 				h.OnResolveMethod(md)
 				h.OnSendHeaders(reqHeaders)
